+++ README.md
@@ -1,0 +1,105 @@
+# Музыкальное приложение для управления плейлистом
+
+Это приложение реализует тестовое задание для управления музыкальным плейлистом с поддержкой CRUD-операций над песнями и функционалом управления воспроизведением: воспроизведение, пауза, следующий и предыдущий трек.
+
+## Эндпоинты
+
+### Операции с плейлистом
+
+1. **Воспроизведение**
+    - **Эндпоинт:** `POST /play`
+    - **Описание:** Запускает воспроизведение текущей песни.
+
+2. **Пауза**
+    - **Эндпоинт:** `POST /pause`
+    - **Описание:** Приостанавливает текущее воспроизведение.
+
+3. **Следующая песня**
+    - **Эндпоинт:** `POST /next`
+    - **Описание:** Переходит к следующей песне.
+
+4. **Предыдущая песня**
+    - **Эндпоинт:** `POST /prev`
+    - **Описание:** Переходит к предыдущей песне.
+
+5. **Текущая песня**
+    - **Эндпоинт:** `GET /current`
+    - **Описание:** Возвращает информацию о текущей воспроизводимой песне.
+
+6. **Получение плейлиста**
+    - **Эндпоинт:** `GET /playlist`
+    - **Описание:** Возвращает список песен в плейлисте.
+ 
+
+### Управление песнями
+
+1. **Добавление песни**
+    - **Эндпоинт:** `POST /songs`
+    - **Описание:** Добавляет новую песню в плейлист.
+    - **Тело запроса:**
+      ```json
+      {
+        "title": "Title",
+        "artist": "Artist",
+        "duration": 100
+      }
+      ```
+
+## Запуск
+
+1. **Клонирование репозитория:**
+   ```bash
+   git clone https://github.com/Velovatik/cloud-go-testtask
+   cd cloud-go-testtask
+   ```
+
+2. **Настройка окружения:**
+    - При необходимости обновите файл конфигурации `config/local.yaml`.
+
+3. **Запуск приложения:**
+    - Выполните скрипт `start.sh`:
+      ```bash
+      ./start.sh
+      ```
+
+   Этот скрипт:
+    - Запускает сервис базы данных.
+    - Применяет миграции базы данных.
+    - Собирает и запускает приложение.
+    - После запуска позволяет просматривать логи приложения
+
+## Подробности реализации
+
+- **Миграции базы данных:** Приложение использует `pressly/goose` для управления миграциями. Миграции находятся в папке `migrations/`.
+- **Обработка конкурентности:** Операции с плейлистом используют `sync.RWMutex` для обеспечения потокобезопасности.
+- **Архитектура:** Код структурирован с разделением на entities, use-cases, repository и delivery.
+- В качестве базы данных используется PostgreSQL внутри docker-compose
+
+## Примеры использования (curl-запросы)
+
+1. **Добавление песни:**
+   ```bash
+   curl -X POST http://localhost:8082/songs -d '{"title": "Imagine", "artist": "John Lennon", "duration": 183}' -H "Content-Type: application/json"
+   ```
+
+2. **Воспроизведение плейлиста:**
+   ```bash
+   curl -X POST http://localhost:8082/play
+   ```
+
+3. **Пауза воспроизведения:**
+   ```bash
+   curl -X POST http://localhost:8082/pause
+   ```
+
+4. **Получение текущей песни:**
+   ```bash
+   curl -X GET http://localhost:8082/current
+   ```
+
+5. **Получение плейлиста:**
+   ```bash
+   curl -X GET http://localhost:8082/playlist
+   ```
+
+P.S. В repository и entity реализовал часть оставшихся CRUD-операций, но не успел соединить их с use-case и хендлерами
