+++ internal/entity/entity.go
@@ -3,5 +3,6 @@ package entity
 import "errors"
 
 var (
-	ErrNullEntity = errors.New("cannot set entity to nil")
+	ErrNullEntity   = errors.New("cannot set entity to nil")
+	ErrSongNotFound = errors.New("song not found")
 )
